% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.r
\name{db_connection}
\alias{db_connection}
\title{Reads a TADM mdb file and extracts the TADM info into a table}
\usage{
db_connection(database_path)
}
\arguments{
\item{database_path}{pathway to the mdb TADM file}
}
\value{
a table containing all the information of the TADM mdb file
}
\description{
This function Reads a TADM mdb file and extracts the TADM info into a table
}
