% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_increments.r
\name{get_n_increments}
\alias{get_n_increments}
\title{Creates a vector of specified length with increments of 10 units.}
\usage{
get_n_increments(n)
}
\arguments{
\item{n}{number of values to create}
}
\value{
a vector of length n with increments of 10 units
}
\description{
This function loads an integer value (n) and retrieves a vector of length n
with values starting at zero and increments of 10 units.
This function is used to create the Time values of a TADM table.
}
