% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TADM_curves.r
\name{plot_TADM_curves}
\alias{plot_TADM_curves}
\title{Reads the main table and plots all the TADM curves.}
\usage{
plot_TADM_curves(main_table, bln_runanalysis, str_testname)
}
\arguments{
\item{main_table}{main table created with db_connection}

\item{bln_runanalysis}{analysis mode, one of 1, 2 or 3 (1=browser; 2=pdf report; 3=both)}

\item{str_testname}{test name for the pdf report}
}
\value{
TADM curves in browser and/or pdf report
}
\description{
This function reads the main table created with db_connection and
plot all the TADM curves into the browser and/or pdf report.
}
